improved minor changes
