+++ pypesto/C.py
@@ -265,6 +265,9 @@ EXPONENTIAL_DECAY = (
     "exponential_decay"  # temperature schedule for parallel tempering
 )
 BETA_DECAY = "beta_decay"  # temperature schedule for parallel tempering
+TRAPEZOID = "trapezoid"  # method to compute log evidence
+SIMPSON = "simpson"  # method to compute log evidence
+STEPPINGSTONE = "steppingstone"  # method to compute log evidence
 
 ###############################################################################
 # PREDICT
