+++ test/run_notebook.sh
@@ -38,7 +38,7 @@ nbs_2=(
   'sampler_study.ipynb'
   'sampling_diagnostics.ipynb'
   'model_selection.ipynb'
-  'bayes_factors.ipynb'
+  'model_evidence_and_bayes_factors.ipynb'
 )
 
 # All tested notebooks
